requirements.txt
Flask
librosa
numpy
soundfile
tensorflow
tensorflow-hub
pandas
matplotlib
tensorflow-io
ipython
speechbrain
opencv-python
ultralytics
torch
transformers
Pillow
translators
pytesseract
moviepy
scikit-learn
python-docx
pymorphy3
nltk
llama-cpp-python
git+https://github.com/jianfch/stable-ts.git